expr/cmp: fix type size
Since nftables now uses nft_rule_expr_get_u32(), the internal size must
also be a uint32_t.

Fixes complete breakage of any cmp or meta expression.

+++ src/expr/cmp.c
@@ -26,8 +26,8 @@
 
 struct nft_expr_cmp {
 	union nft_data_reg	data;
-	uint8_t			sreg;	/* enum nft_registers */
-	uint8_t			op;	/* enum nft_cmp_ops */
+	uint32_t		sreg;	/* enum nft_registers */
+	uint32_t		op;	/* enum nft_cmp_ops */
 };
 
 static int
